+++ blocklets/image-bin/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.13.17 (August 28, 2025)
 
-- chore: use aigne hub image instand of call api
+- chore: use aigne hub image instead of call api
 
 ## 0.13.16 (August 27, 2025)